refactor: one method controlls all events
