+++ omni-relayer/redis-utility.sh
@@ -43,31 +43,11 @@ case "$1" in
   get_solana_last_processed_signature)
 	get_last_processed "Solana" "SOLANA_LAST_PROCESSED_SIGNATURE"
     ;;
+  get_events)
+	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "events" | sed -n '1~2p'
+    ;;
   get_solana_events)
 	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "solana_events" | sed -n '1~2p'
-    ;;
-  get_init_transfers)
-	echo "Init transfers on Near:"
-	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "near_init_transfer_queue" | sed -n 'n;p' | sed G
-	echo "Init transfers on Evm:"
-	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "evm_init_tranfer_events" | sed -n 'n;p' | sed G
-	echo "Init transfers on Solana:"
-	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "solana_init_transfer_events" | sed -n 'n;p' | sed G
-    ;;
-  get_near_init_transfer_events)
-	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "near_init_transfer_queue" | sed -n 'n;p' | sed G
-    ;;
-  get_evm_init_transfer_events)
-	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "evm_init_tranfer_events" | sed -n 'n;p' | sed G
-    ;;
-  get_solana_init_transfer_events)
-	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "solana_init_transfer_events" | sed -n 'n;p' | sed G
-    ;;
-  get_sign_events)
-	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "near_sign_transfer_events" | sed -n 'n;p' | sed G
-    ;;
-  get_finalized_transfers)
-	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "finalized_transfers" | sed -n 'n;p' | sed G
     ;;
   get_stuck_transfers)
 	redis-cli -h "$REDIS_HOST" -p "$REDIS_PORT" hgetall "stuck_transfers" | sed -n 'n;p' | sed G
@@ -85,13 +65,8 @@ case "$1" in
     echo "  get_arb_last_processed_block        - Retrieve the last processed block for Arbitrum"
     echo "  get_evm_last_processed_block        - Retrieve the last processed blocks for all EVM chains"
     echo "  get_solana_last_processed_signature - Retrieve the last processed signature for Solana"
+    echo "  get_events                          - Retrieve all events related to Near"
     echo "  get_solana_events                   - Retrieve all event keys for Solana"
-    echo "  get_init_transfers                  - Retrieve all init transfer events for all chains"
-    echo "  get_near_init_transfer_events       - Retrieve init transfer events for Near"
-    echo "  get_evm_init_transfer_events        - Retrieve init transfer events for EVM"
-    echo "  get_solana_init_transfer_events     - Retrieve init transfer events for Solana"
-    echo "  get_sign_events                     - Retrieve all sign events for Near"
-    echo "  get_finalized_transfers             - Retrieve finalized transfer events"
     echo "  get_stuck_transfers                 - Retrieve stuck transfer events"
     exit 1
     ;;
