+++ omni-relayer/src/main.rs
@@ -75,31 +75,6 @@ async fn main() -> Result<()> {
             .await
         }
     }));
-    handles.push(tokio::spawn({
-        #[cfg(not(feature = "disable_fee_check"))]
-        let config = config.clone();
-        let redis_client = redis_client.clone();
-        let connector = connector.clone();
-        let jsonrpc_client = jsonrpc_client.clone();
-        let near_nonce = near_nonce.clone();
-        async move {
-            workers::near::sign_transfer(
-                #[cfg(not(feature = "disable_fee_check"))]
-                config,
-                redis_client,
-                connector,
-                jsonrpc_client,
-                near_nonce,
-            )
-            .await
-        }
-    }));
-    handles.push(tokio::spawn({
-        let redis_client = redis_client.clone();
-        let connector = connector.clone();
-        let evm_nonces = evm_nonces.clone();
-        async move { workers::near::finalize_transfer(redis_client, connector, evm_nonces).await }
-    }));
 
     if config.eth.is_some() {
         handles.push(tokio::spawn({
@@ -158,37 +133,7 @@ async fn main() -> Result<()> {
         handles.push(tokio::spawn({
             let config = config.clone();
             let redis_client = redis_client.clone();
-            async move { workers::solana::process_signature(config, redis_client).await }
-        }));
-        handles.push(tokio::spawn({
-            let config = config.clone();
-            let redis_client = redis_client.clone();
-            let connector = connector.clone();
-            let near_nonce = near_nonce.clone();
-            async move {
-                workers::solana::finalize_transfer(config, redis_client, connector, near_nonce)
-                    .await
-            }
-        }));
-    }
-
-    if config.eth.is_some() || config.base.is_some() || config.arb.is_some() {
-        handles.push(tokio::spawn({
-            let config = config.clone();
-            let redis_client = redis_client.clone();
-            let connector = connector.clone();
-            let jsonrpc_client = jsonrpc_client.clone();
-            let near_nonce = near_nonce.clone();
-            async move {
-                workers::evm::finalize_transfer(
-                    config,
-                    redis_client,
-                    connector,
-                    jsonrpc_client,
-                    near_nonce,
-                )
-                .await
-            }
+            async move { startup::solana::process_signature(config, redis_client).await }
         }));
     }
 
@@ -198,21 +143,18 @@ async fn main() -> Result<()> {
         let connector = connector.clone();
         let jsonrpc_client = jsonrpc_client.clone();
         let near_nonce = near_nonce.clone();
-        async move {
-            workers::near::claim_fee(config, redis_client, connector, jsonrpc_client, near_nonce)
-                .await
-        }
-    }));
+        let evm_nonces = evm_nonces.clone();
 
-    handles.push(tokio::spawn({
-        let config = config.clone();
-        let redis_client = redis_client.clone();
-        let connector = connector.clone();
-        let jsonrpc_client = jsonrpc_client.clone();
-        let near_nonce = near_nonce.clone();
         async move {
-            workers::near::bind_token(config, redis_client, connector, jsonrpc_client, near_nonce)
-                .await
+            workers::process_events(
+                config,
+                redis_client,
+                connector,
+                jsonrpc_client,
+                near_nonce,
+                evm_nonces,
+            )
+            .await
         }
     }));
