+++ omni-relayer/src/utils/redis.rs
@@ -2,18 +2,10 @@ use log::warn;
 use omni_types::ChainKind;
 use redis::{aio::MultiplexedConnection, AsyncCommands};
 
-pub const DEPLOY_TOKEN_EVENTS: &str = "deploy_token_events";
-
-pub const NEAR_INIT_TRANSFER_QUEUE: &str = "near_init_transfer_queue";
-pub const NEAR_SIGN_TRANSFER_EVENTS: &str = "near_sign_transfer_events";
-
-pub const EVM_INIT_TRANSFER_EVENTS: &str = "evm_init_tranfer_events";
-
+pub const EVENTS: &str = "events";
 pub const SOLANA_EVENTS: &str = "solana_events";
-pub const SOLANA_INIT_TRANSFER_EVENTS: &str = "solana_init_transfer_events";
 
-pub const FINALIZED_TRANSFERS: &str = "finalized_transfers";
-pub const STUCK_TRANSFERS: &str = "stuck_transfers";
+pub const STUCK_EVENTS: &str = "stuck_events";
 
 pub const KEEP_INSUFFICIENT_FEE_TRANSFERS_FOR: i64 = 60 * 60 * 24 * 14; // 14 days
 pub const CHECK_INSUFFICIENT_FEE_TRANSFERS_EVERY_SECS: i64 = 60 * 30; // 30 minutes
